chore: refactoring eventBus for use in shared
