+++ frontend/libs/portal/components/dialog/LicenseRuleDialog.vue
@@ -6,7 +6,7 @@
 import {DialogLicenseRuleConfig} from '@disclosure-portal/components/dialog/DialogConfigs';
 import {ErrorDialogInterface} from '@disclosure-portal/components/dialog/DialogInterfaces';
 import ErrorDialog from '@disclosure-portal/components/dialog/ErrorDialog.vue';
-import ErrorDialogConfig from '@disclosure-portal/model/ErrorDialogConfig';
+import ErrorDialogConfig from '@shared/types/ErrorDialogConfig';
 import {LicenseRuleRequest} from '@disclosure-portal/model/LicenseRule';
 import {ComponentLicenses} from '@disclosure-portal/model/Project';
 import {ComponentInfoSlim} from '@disclosure-portal/model/VersionDetails';
