+++ frontend/libs/portal/components/dialog/project/RequestReview.vue
@@ -10,7 +10,7 @@ import {ComponentStats, SpdxFile, VersionSlim} from '@disclosure-portal/model/Ve
 import profileService from '@disclosure-portal/services/profile';
 import projectService from '@disclosure-portal/services/projects';
 import versionService from '@disclosure-portal/services/version';
-import {useIdleStore} from '@disclosure-portal/stores/idle.store';
+import {useIdleStore} from '@shared/stores/idle.store';
 import {useProjectStore} from '@disclosure-portal/stores/project.store';
 import {useSbomStore} from '@disclosure-portal/stores/sbom.store';
 import useRules from '@disclosure-portal/utils/Rules';
