+++ frontend/libs/portal/views/projects/ProjectsDetail.vue
@@ -6,7 +6,7 @@
 import {usePageTitle} from '@disclosure-portal/composables/usePageTitle';
 import {ProjectSubscriptions} from '@disclosure-portal/model/Project';
 import {useAppStore} from '@disclosure-portal/stores/app';
-import {useIdleStore} from '@disclosure-portal/stores/idle.store';
+import {useIdleStore} from '@shared/stores/idle.store';
 import {useProjectStore} from '@disclosure-portal/stores/project.store';
 import {useBreadcrumbsStore} from '@shared/stores/breadcrumbs.store';
 import _ from 'lodash';
