+++ frontend/libs/shared/components/widgets/Idle.vue
@@ -3,7 +3,7 @@
 <!-- SPDX-License-Identifier: Apache-2.0 -->
 
 <script setup lang="ts">
-import {useIdleStore} from '@disclosure-portal/stores/idle.store';
+import {useIdleStore} from '@shared/stores/idle.store';
 import {SelfBuildingSquareSpinner} from 'epic-spinners';
 import {VProgressCircular} from 'vuetify/components';
