+++ frontend/libs/shared/types/DHTTPError.ts
