+++ frontend/libs/shared/types/ErrorDialogConfig.ts
