+++ frontend/libs/shared/types/IdleInfo.ts
