+++ frontend/libs/shared/types/eventBus.d.ts
@@ -2,21 +2,16 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
-import DHTTPError from '@disclosure-portal/model/DHTTPError';
-import ErrorDialogConfig from '@disclosure-portal/model/ErrorDialogConfig';
-import IdleInfo, {INotificationMeta} from '@disclosure-portal/model/IdleInfo';
-import mitt from 'mitt';
+import DHTTPError from '@shared/types/DHTTPError';
+import ErrorDialogConfig from '@shared/types/ErrorDialogConfig';
+import IdleInfo, {INotificationMeta} from '@shared/types/IdleInfo';
 
-type Events = {
+export type Events = {
   'show-snackbar': {message: string; timeout?: number; level: string};
   'on-api-error': DHTTPError;
   'on-error': {error: ErrorDialogConfig};
   'on-idle': {idle: IdleInfo};
-  'window-resize': {};
+  'window-resize': unknown;
   'set-notification': {config: INotificationMeta};
   'tab-change': {tabIndex: number};
 };
-
-const eventBus = mitt<Events>();
-
-export default eventBus;
